Simple Stateful Scheduler (SSS)
This commit implements SSS. Essentially we want the system to "just
work", so nothing fancy going on here. In order to make sure that
messages to stateful actors reachthe correct invoker and don't start a
new invocation on a different invoker, all activations to a stateful
action must be scheduled to a predictable invoker.

This means that if an invoker is known to host a particular actor and it
is in any state other than "Offline", we should send the stateful
activations to it. That's all that this scheduler implements